+++ api-core-common/src/main/java/gov/ca/cwds/tracelog/delegate/DelegateTraceLogRecordAccessDao.java
@@ -9,6 +9,7 @@ import gov.ca.cwds.tracelog.simple.SimpleTraceLogRecordAccessDao;
 
 public class DelegateTraceLogRecordAccessDao implements TraceLogRecordAccessDao {
 
+  @Inject
   private TraceLogRecordAccessDao dao = new SimpleTraceLogRecordAccessDao();
 
   @Override
@@ -20,7 +21,6 @@ public class DelegateTraceLogRecordAccessDao implements TraceLogRecordAccessDao 
     return dao;
   }
 
-  @Inject
   public void setDao(TraceLogRecordAccessDao dao) {
     this.dao = dao;
   }
