+++ legacy-data-access/src/main/java/gov/ca/cwds/tracelog/dao/TraceLogRecordAccessDaoImpl.java
@@ -3,7 +3,10 @@ package gov.ca.cwds.tracelog.dao;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 
+import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 
@@ -15,6 +18,8 @@ import gov.ca.cwds.tracelog.entity.TraceLogClientViewLog;
 public class TraceLogRecordAccessDaoImpl extends BaseDaoImpl<TraceLogClientViewLog>
     implements TraceLogRecordAccessDao {
 
+  private static final Logger LOGGER = LoggerFactory.getLogger(TraceLogRecordAccessDaoImpl.class);
+
   @Inject
   public TraceLogRecordAccessDaoImpl(@CmsSessionFactory SessionFactory sessionFactory) {
     super(sessionFactory);
@@ -22,7 +27,10 @@ public class TraceLogRecordAccessDaoImpl extends BaseDaoImpl<TraceLogClientViewL
 
   @Override
   public void logRecordAccess(String userId, LocalDateTime moment, String id, String entityType) {
-    create(new TraceLogClientViewLog(userId, Timestamp.valueOf(moment), id, entityType));
+    LOGGER.info("Trace Log, log access: user: {}, entity: {}, id: {}", userId, entityType, id);
+    try (final Session session = getSessionFactory().openSession()) {
+      create(new TraceLogClientViewLog(userId, Timestamp.valueOf(moment), id, entityType));
+    }
   }
 
 }
